nlp_yelp_reviews/requirements.txt
asgiref                3.5.2
backports.zoneinfo     0.2.1
click                  8.1.3
colorama               0.4.5
contourpy              1.0.5
cycler                 0.11.0
Django                 4.1.3
django-cors-headers    3.13.0
djangorestframework    3.14.0
fonttools              4.37.4
greenlet               2.0.0
joblib                 1.2.0
kiwisolver             1.4.4
matplotlib             3.6.1
mysql-connector        2.2.9
mysql-connector-python 8.0.31
mysqlclient            2.1.1
nltk                   3.7
numpy                  1.23.4
packaging              21.3
pandas                 1.5.0
Pillow                 9.2.0
pip                    21.3.1
plotly                 5.11.0
protobuf               3.20.1
px                     0.1.0
PyMySQL                1.0.2
pyparsing              3.0.9
python-dateutil        2.8.2
pytz                   2022.4
regex                  2022.9.13
scikit-learn           1.1.2
scipy                  1.9.2
seaborn                0.12.0
setuptools             60.2.0
six                    1.16.0
SQLAlchemy             1.4.42
sqlparse               0.4.3
tenacity               8.1.0
textblob               0.17.1
threadpoolctl          3.1.0
tqdm                   4.64.1
tzdata                 2022.6
websocket-client       1.4.2
wheel                  0.37.1
wordcloud              1.8.2.2
